# docs/swagger.yaml
definitions:
...
  handler.OtherOutStockItemReq:
    properties:
      price:
        type: number
      productId:
        type: integer
      quantity:
        type: number
      remark:
        type: string
    required:
    - productId
    - quantity
    type: object
  handler.OtherOutStockSaveReq:
    properties:
      billDate:
        type: string
      complete:
        type: boolean
      counterpart:
        type: string
      deptId:
        type: integer
      handlerId:
        type: integer
      items:
        items:
          $ref: '#/definitions/handler.OtherOutStockItemReq'
        minItems: 1
        type: array
      outType:
        type: string
      remark:
        type: string
      settleUnit:
        type: string
      warehouseId:
        type: integer
    required:
    - billDate
    - items
    - outType
    - warehouseId
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /other-out-stocks:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 单号
        in: query
        name: keyword
        type: string
      - description: 商品名称/编号
        in: query
        name: productKw
        type: string
      - description: 出库仓库
        in: query
        name: warehouseId
        type: integer
      - description: 状态 draft/completed
        in: query
        name: status
        type: string
      - description: 出库类型
        in: query
        name: outType
        type: string
      - description: 录单时间起 YYYY-MM-DD
        in: query
        name: startDate
        type: string
      - description: 录单时间止 YYYY-MM-DD
        in: query
        name: endDate
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 其他出库单列表
      tags:
      - 库存
    post:
      parameters:
      - description: 单据内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.OtherOutStockSaveReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 新增其他出库单（草稿或保存并过账）
      tags:
      - 库存
  /other-out-stocks/{id}:
    delete:
      parameters:
      - description: 单据ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除其他出库单
      tags:
      - 库存
    get:
      parameters:
      - description: 单据ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 其他出库单详情
      tags:
      - 库存
    put:
      parameters:
      - description: 单据ID
        in: path
        name: id
        required: true
        type: integer
      - description: 单据内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.OtherOutStockSaveReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 编辑其他出库单
      tags:
      - 库存
  /other-out-stocks/{id}/complete:
    put:
      parameters:
      - description: 单据ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 其他出库单过账
      tags:
      - 库存
...
swagger: "2.0"